+++ src/main/java/org/bigtows/window/ui/notetree/NoteTree.java
@@ -10,6 +10,7 @@ import org.bigtows.window.ui.notetree.tree.entity.Task;
 import org.bigtows.window.ui.notetree.tree.node.AbstractTaskTreeNode;
 import org.bigtows.window.ui.notetree.tree.node.NoteTreeNode;
 import org.bigtows.window.ui.notetree.tree.node.TaskTreeNode;
+import org.bigtows.window.ui.notetree.tree.transfer.TreeTransferHandler;
 import org.bigtows.window.ui.notetree.utils.ExpandTreeUtils;
 
 import javax.swing.*;
